HSEARCH-3589 Introduce a common superclass for object fields and value fields in the Elasticsearch metamodel
