+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/document/model/impl/ElasticsearchIndexSchemaFieldNode.java
@@ -6,48 +6,27 @@
  */
 package org.hibernate.search.backend.elasticsearch.document.model.impl;
 
-import java.lang.invoke.MethodHandles;
 import java.util.List;
 
-import org.hibernate.search.backend.elasticsearch.gson.impl.JsonAccessor;
-import org.hibernate.search.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.backend.elasticsearch.types.impl.ElasticsearchIndexFieldType;
 import org.hibernate.search.engine.backend.document.model.spi.IndexFieldInclusion;
-import org.hibernate.search.engine.backend.metamodel.IndexObjectFieldDescriptor;
 import org.hibernate.search.engine.backend.metamodel.IndexValueFieldDescriptor;
 import org.hibernate.search.engine.reporting.spi.EventContexts;
-import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 import org.hibernate.search.util.common.reporting.EventContext;
 
-import com.google.gson.JsonElement;
 
-
-public class ElasticsearchIndexSchemaFieldNode<F> implements IndexValueFieldDescriptor {
-
-	private static final Log log = LoggerFactory.make( Log.class, MethodHandles.lookup() );
-
-	private final ElasticsearchIndexSchemaObjectNode parent;
-	private final String absolutePath;
-	private final String relativeName;
-	private final JsonAccessor<JsonElement> relativeAccessor;
-	private final IndexFieldInclusion inclusion;
+public class ElasticsearchIndexSchemaFieldNode<F> extends AbstractElasticsearchIndexSchemaFieldNode
+		implements IndexValueFieldDescriptor {
 
 	private final List<String> nestedPathHierarchy;
-
-	private final boolean multiValued;
 
 	private final ElasticsearchIndexFieldType<F> type;
 
 	public ElasticsearchIndexSchemaFieldNode(ElasticsearchIndexSchemaObjectNode parent, String relativeFieldName,
 			IndexFieldInclusion inclusion, boolean multiValued, ElasticsearchIndexFieldType<F> type) {
-		this.parent = parent;
-		this.absolutePath = parent.absolutePath( relativeFieldName );
-		this.relativeName = relativeFieldName;
-		this.relativeAccessor = JsonAccessor.root().property( relativeFieldName );
-		this.inclusion = inclusion;
+		super( parent, relativeFieldName, inclusion, multiValued );
 		this.nestedPathHierarchy = parent.getNestedPathHierarchy();
 		this.type = type;
-		this.multiValued = multiValued;
 	}
 
 	@Override
@@ -61,36 +40,13 @@ public class ElasticsearchIndexSchemaFieldNode<F> implements IndexValueFieldDesc
 	}
 
 	@Override
-	public IndexObjectFieldDescriptor toObjectField() {
+	public ElasticsearchIndexSchemaObjectFieldNode toObjectField() {
 		throw log.invalidIndexElementTypeValueFieldIsNotObjectField( absolutePath );
 	}
 
 	@Override
-	public IndexValueFieldDescriptor toValueField() {
+	public ElasticsearchIndexSchemaFieldNode<F> toValueField() {
 		return this;
-	}
-
-	@Override
-	public ElasticsearchIndexSchemaObjectNode parent() {
-		return parent;
-	}
-
-	@Override
-	public String absolutePath() {
-		return absolutePath;
-	}
-
-	@Override
-	public String relativeName() {
-		return relativeName;
-	}
-
-	public JsonAccessor<JsonElement> getRelativeAccessor() {
-		return relativeAccessor;
-	}
-
-	public IndexFieldInclusion getInclusion() {
-		return inclusion;
 	}
 
 	public String getNestedPath() {
@@ -101,11 +57,6 @@ public class ElasticsearchIndexSchemaFieldNode<F> implements IndexValueFieldDesc
 
 	public List<String> getNestedPathHierarchy() {
 		return nestedPathHierarchy;
-	}
-
-	@Override
-	public boolean isMultiValued() {
-		return multiValued;
 	}
 
 	@Override
